Hash for known beacons

+++ src/core/utility.c
@@ -612,7 +612,7 @@ void print_and_free_error(GError *error)
 /*
    Read byte array and compute hash from GVariant
 */
-unsigned char *read_byte_array(GVariant *s_value, int *actualLength, uint32_t *hash)
+unsigned char *read_byte_array(GVariant *s_value, int *actualLength, uint16_t *hash)
 {
     unsigned char byteArray[2048];
     int len = 0;
